% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avoidances}
\alias{avoidances}
\title{avoidances}
\format{
A data frame with 349 rows and 5 variables:
\describe{
\item{\code{ID}}{double Participant ID code}
\item{\code{foodavoid}}{character Self-reported foods avoided during pregnancy}
\item{\code{whotaught}}{character Who told the participant to avoid that food}
\item{\code{consequence}}{character Self-reported consequences of eating the food during pregnancy}
\item{\code{kembara explained}}{character A notes column on the participants understanding of local illness category kembara}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
avoidances
}
\description{
Learned food avoidances during pregnancy in Mysore, India, 2015
}
\keyword{datasets}
