% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycars.R
\docType{data}
\name{mycars}
\alias{mycars}
\title{Cars}
\format{An object of class \code{data.frame} with 32 rows and 11 columns.}
\usage{
mycars
}
\description{
A really great cars data set.
}
\keyword{datasets}
